backend/requirements.txt
# ── Core Framework ──────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.16.2
pydantic==2.9.2
pydantic-settings==2.5.2

# ── Database ────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.10

# ── Data/ML ─────────────────────────────────────────────────────────────────
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
xgboost==2.1.1
imbalanced-learn==0.13.0
joblib==1.4.2
catboost>=1.2.0
google-genai>=1.0.0


# ── Clustering/Embeddings ───────────────────────────────────────────────────
umap-learn==0.5.7
sentence-transformers==3.1.1

# ── LLM ─────────────────────────────────────────────────────────────────────
httpx==0.27.2
groq>=0.9.0

# ── Utilities ───────────────────────────────────────────────────────────────
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                